+++ check/mode-common.c
@@ -403,8 +403,8 @@ int insert_inode_item(struct btrfs_trans_handle *trans,
 
 	warning("root %llu inode %llu recreating inode item, this may "
 		"be incomplete, please check permissions and content after "
-		"the fsck completes.\n", (unsigned long long)root->objectid,
-		(unsigned long long)ino);
+		"the fsck completes.\n",
+		root->objectid, ino);
 
 	return 0;
 }
